modify the instantiation section according to the review
