+++ instantiation/__main__.py
@@ -1,0 +1,7 @@
+# Copyright (c) Microsoft Corporation.
+# Licensed under the MIT License.
+from instantiation import instantiation
+
+if __name__ == "__main__":
+    # Execute the main script
+    instantiation.main()
