+++ instantiation/config/config.py
@@ -7,7 +7,7 @@ from ufo.config.config import Config
 class Config(Config):
     _instance = None
 
-    def __init__(self, config_path="instantiation/controller/config/"):
+    def __init__(self, config_path="instantiation/config/"):
         self.config_data = self.load_config(config_path)
 
     @staticmethod
@@ -22,7 +22,7 @@ class Config(Config):
         return Config._instance
 
     def optimize_configs(self, configs):
-        self.update_api_base(configs, "ACTION_PREFILL_AGENT")
+        self.update_api_base(configs, "PREFILL_AGENT")
         self.update_api_base(configs, "FILTER_AGENT")
 
         return configs
