+++ instantiation/config/config.yaml.template
@@ -1,0 +1,43 @@
+# You will configure for the prefill agent and filter agent individualy.
+# Prefill agent is used to prefill the task.
+# Filter agent is to evaluate the prefill quality.
+
+PREFILL_AGENT: {
+  VISUAL_MODE: True, # Whether to use the visual mode
+
+  API_TYPE: "azure_ad" , # The API type, "openai" for the OpenAI API, "aoai" for the AOAI API, 'azure_ad' for the ad authority of the AOAI API.  
+  API_BASE: "https://cloudgpt-openai.azure-api.net/", # The the OpenAI API endpoint, "https://api.openai.com/v1/chat/completions" for the OpenAI API. As for the AAD, it should be your endpoints.
+  API_KEY: "YOUR_API_KEY",  # The OpenAI API key
+  API_VERSION: "2024-02-15-preview", # "2024-02-15-preview" by default
+  API_MODEL: "gpt-4o-20240513",  # The only OpenAI model by now that accepts visual input
+
+  ###For the AOAI
+  API_DEPLOYMENT_ID: "gpt-4-0125-preview", # The deployment id for the AOAI API
+  ### For Azure_AD
+  AAD_TENANT_ID: "YOUR_AAD_ID", # Set the value to your tenant id for the llm model
+  AAD_API_SCOPE: "openai", # Set the value to your scope for the llm model
+  AAD_API_SCOPE_BASE: "YOUR_AAD_API_SCOPE_BASE"  # Set the value to your scope base for the llm model, whose format is API://YOUR_SCOPE_BASE, and the only need is the YOUR_SCOPE_BASE
+}
+
+FILTER_AGENT: {
+  VISUAL_MODE: False, # Whether to use the visual mode
+
+  API_TYPE: "azure_ad" , # The API type, "openai" for the OpenAI API, "aoai" for the Azure OpenAI.  
+  API_BASE: "https://cloudgpt-openai.azure-api.net/", # The the OpenAI API endpoint, "https://api.openai.com/v1/chat/completions" for the OpenAI API. As for the aoai, it should be https://{your-resource-name}.openai.azure.com
+  API_KEY: "YOUR_API_KEY",  # The aoai API key
+  API_VERSION: "2024-04-01-preview", # "2024-02-15-preview" by default
+  API_MODEL: "gpt-4o-20240513",  # The only OpenAI model by now that accepts visual input
+  API_DEPLOYMENT_ID: "gpt-4o-20240513-preview", # The deployment id for the AOAI API
+  
+  ### For Azure_AD
+  AAD_TENANT_ID: "YOUR_AAD_ID",
+  AAD_API_SCOPE: "openai", #"openai"
+  AAD_API_SCOPE_BASE: "YOUR_AAD_API_SCOPE_BASE", #API://YOUR_SCOPE_BASE
+}
+
+# For parameters
+MAX_TOKENS: 2000  # The max token limit for the response completion
+MAX_RETRY: 3  # The max retry limit for the response completion
+TEMPERATURE: 0.0  # The temperature of the model: the lower the value, the more consistent the output of the model
+TOP_P: 0.0  # The top_p of the model: the lower the value, the more conservative the output of the model
+TIMEOUT: 60  # The call timeout(s), default is 10 minss
