+++ instantiation/config/config_dev.yaml
@@ -1,0 +1,30 @@
+version: 0.1
+
+AOAI_DEPLOYMENT: "gpt-4-visual-preview"  # Your AOAI deployment if apply
+API_VERSION: "2024-02-15-preview" # "2024-02-15-preview" by default.
+OPENAI_API_MODEL: "gpt-4-0125-preview"  # The only OpenAI model by now that accepts visual input
+
+CONTROL_BACKEND: "uia"  # The backend for control action
+CONTROL_LIST: ["Button", "Edit", "TabItem", "Document", "ListItem", "MenuItem", "ScrollBar", "TreeItem", "Hyperlink", "ComboBox", "RadioButton", "DataItem", "Spinner"] 
+PRINT_LOG: False  # Whether to print the log  
+LOG_LEVEL: "DEBUG"  # The log level
+
+PREFILL_PROMPT: "instantiation/controller/prompts/{mode}/prefill.yaml"  # The prompt for the action prefill
+FILTER_PROMPT: "instantiation/controller/prompts/{mode}/filter.yaml"  # The prompt for the filter
+PREFILL_EXAMPLE_PROMPT: "instantiation/controller/prompts/{mode}/prefill_example.yaml"  # The prompt for the action prefill example
+API_PROMPT: "instantiation/controller/prompts/{mode}/api.yaml"  # The prompt for the API
+
+# Exploration Configuration
+TASKS_HUB: "instantiation/tasks"  # The tasks hub for the exploration
+TEMPLATE_PATH: "instantiation/templates"  # The template path for the exploration
+
+# For control filtering
+CONTROL_FILTER_TYPE: []  # The list of control filter type, support 'TEXT', 'SEMANTIC', 'ICON'
+CONTROL_FILTER_MODEL_SEMANTIC_NAME: "all-MiniLM-L6-v2"  # The control filter model name of semantic similarity
+CONTROL_EMBEDDING_CACHE_PATH: "instantiation/cache/"  # The cache path for the control filter
+CONTROL_FILTER_TOP_K_PLAN: 2  # The control filter effect on top k plans from UFO, default is 2
+
+# log path
+LOG_PATH: "instantiation/logs/{task}"
+PREFILL_LOG_PATH: "instantiation/logs/{task}/prefill/"
+FILTER_LOG_PATH: "instantiation/logs/{task}/filter/"
